fix x64 problem in smapi.cpp and APEDecompress.cpp remove mpclib , because it is not used anymore ( was replaced by libmpcdec a while ago)
converted windows vcproj and sln to vs 2008. adjust manifest to contain vistas trustinfo
vs 2005 projects have been renamed to xx_vs2005.vcproj/sln so vs2005 can be still be used ( needed for windows 98 build)

+++ trunk/wxMusik/MUSIKEngine/3rd-Party/MAC/MACLib/APEDecompress.cpp
@@ -23,7 +23,7 @@ CAPEDecompress::CAPEDecompress(int * pErrorCode, CAPEInfo * pAPEInfo, int nStart
     }
 
     // get format information
-    GetInfo(APE_INFO_WAVEFORMATEX, (int) &m_wfeInput);
+    GetInfo(APE_INFO_WAVEFORMATEX, (intptr_t) &m_wfeInput);
     m_nBlockAlign = GetInfo(APE_INFO_BLOCK_ALIGN);
 
     // initialize other stuff
